Move Scala 3 logstage macros to macros subpackage
